// src/common/models/course.ts
import { Chevaux } from "./chevaux.js";
import { JockeyDriver } from "./jockeyDriver.js";
import { Multi } from "./multi.js";
import { Paris } from "./paris.js";
import { Quarte } from "./quarte.js";
import { Quinte } from "./quinte.js";
import { Tierce } from "./tierce.js";

export interface Course {
    cached: boolean;
    departImminent: boolean;
    arriveeDefinitive: boolean;
    timezoneOffset: number;
    numReunion: number;
    numExterneReunion: number;
    numOrdre: number;
    numExterne: number;
    heureDepart: number;
    libelle: string;
    libelleCourt: string;
    montantPrix: number;
    parcours: string;
    distance: number;
    distanceUnit: string;
    corde: string;
    discipline: string;
    specialite: string;
    categorieParticularite: string;
    conditionSexe: string;
    nombreDeclaresPartants: number;
    grandPrixNationalTrot: boolean;
    numSocieteMere: number;
    pariMultiCourses: boolean;
    pariSpecial: boolean;
    montantTotalOffert: number;
    montantOffert1er: number;
    montantOffert2eme: number;
    montantOffert3eme: number;
    montantOffert4eme: number;
    montantOffert5eme: number;
    conditions: string;
    numCourseDedoublee: number;
    paris?: Paris[];
    chevaux?: Chevaux[];
    tierce?: Tierce | null;
    quarte?: Quarte | null;
    quinte?: Quinte | null;
    multi?: Multi | null;
    jockeyDriver?: JockeyDriver[];
    id_reunion?: number;
}